Modularise code
Move related segments out of bufferline into their own modules
