+++ lua/bufferline/constants.lua
@@ -1,0 +1,33 @@
+local M = {}
+---------------------------------------------------------------------------//
+-- Constants
+---------------------------------------------------------------------------//
+M.padding = " "
+
+M.superscript_numbers = {
+  [0] = "⁰",
+  [1] = "¹",
+  [2] = "²",
+  [3] = "³",
+  [4] = "⁴",
+  [5] = "⁵",
+  [6] = "⁶",
+  [7] = "⁷",
+  [8] = "⁸",
+  [9] = "⁹",
+  [10] = "¹⁰",
+  [11] = "¹¹",
+  [12] = "¹²",
+  [13] = "¹³",
+  [14] = "¹⁴",
+  [15] = "¹⁵",
+  [16] = "¹⁶",
+  [17] = "¹⁷",
+  [18] = "¹⁸",
+  [19] = "¹⁹",
+  [20] = "²⁰"
+}
+
+M.letters = "abcdefghijklmopqrstuvwxyzABCDEFGHIJKLMOPQRSTUVWXYZ"
+
+return M
