+++ lua/bufferline/tabs.lua
@@ -1,0 +1,46 @@
+local highlights = require "bufferline/highlights"
+local constants = require "bufferline/constants"
+
+local M = {}
+
+local strwidth = vim.fn.strwidth
+local padding = constants.padding
+
+local function tab_click_component(num)
+  return "%" .. num .. "T"
+end
+
+local function render(tab, is_active, style)
+  local hl = is_active and highlights.tab_selected or highlights.tab
+  local separator_hl =
+    is_active and highlights.selected_separator or highlights.separator
+  local separator_component = style == "thick" and "▐" or "▕"
+  local separator = separator_hl .. separator_component
+  local name = padding .. padding .. tab.tabnr .. padding
+  local length = strwidth(name) + strwidth(separator_component)
+  return hl .. tab_click_component(tab.tabnr) .. name .. separator, length
+end
+
+--- @param style string
+function M.get(style)
+  local all_tabs = {}
+  local tabs = vim.fn.gettabinfo()
+  local current_tab = vim.fn.tabpagenr()
+
+  -- use ordinals to ensure contiguous keys in the table i.e. an array
+  -- rather than an object
+  -- GOOD = {1: thing, 2: thing} BAD: {1: thing, [5]: thing}
+  for i, tab in ipairs(tabs) do
+    local is_active_tab = current_tab == tab.tabnr
+    local component, length = render(tab, is_active_tab, style)
+    all_tabs[i] = {
+      component = component,
+      length = length,
+      id = tab.tabnr,
+      windows = tab.windows
+    }
+  end
+  return all_tabs
+end
+
+return M
